// src/app/model/registro-evento.model.ts
export interface AsisRegEventBol {
  nombreAsistente: string;
  apellidoPaternoAsistente: string;
  apellidoMaternoAsistente: string;
  correoAsistente: string;
  tipoAsistente: string;
  semestre: number;
  telefono: number;
  matricula: number;
  evento_ID: number;
  fecha_Registro: string;
  estadoPago: string;
  asiento: string;
}